Unified callback: async calls in init stack
Problem: If the application makes an async call in the initial
stack (as done in the Micro '15 benchmarks code with an IO
to an output file), then we didn't have a parent. We assumed async
calls (to the threadpool ) would always have a parent.

Solution: When submitting work to the threadpool, if there is no
active CBN then we assume we're in the initial stack. Add a special
function to return a dummy CBN in this case.

This commit also changes the graphviz tree output to print directly
to a file for ease of visualization.