list: thread safe and unit test
1. Add a _lock field and static list__lock, list__unlock routines.
Acquire and release the lock in all(?) relevant place.
The list class should now be thread safe, though I have not
explicitly tested this.

2. Add a unit test for the list class.

3. list_remove now requires that you provide the list*.
This is needful lest the n_elts field of a list be corrupted,
and also to ensure thread safety. If this becomes infeasible,
consider tracking the list* CONTAINER in the list_elem so that
the list_elem itself knows where the list is. This uses an extra
pointer in every element, but efficiency is not of particular
concern.