uv-common: Fix compiler warnings, call class UTs
Call list_UT and map_UT the first time through.

This ensures that bugs arising from subsequent changes to those
classes are caught early. This assumes of course that changes
in the classes are matched by changes in the UT functions.

This is debug code and can be removed in production.