timer: get parent; maintain tree in repeat timers
1. Was not correctly setting the parent of a timer. TODO TEST THIS
2. In a repeating timer, each new CB is the child of the previous CB.

NB This makes use of the declaration of a new variable
'struct callback_node blah = ' inside every INVOKE_CALLBACK
as added in e36d94ec9128. This might not have been the best idea.
A better idea would be to set a user-defined pointer that is NULL
in most cases.

However, since e36d94ec9128 compiles, I guess all of the INVOKE_*
things are inside code blocks

...
{
  INVOKE_*
}

...
{
  INVOKE_*
}

so it doesn't matter.