libuv: Add and use API uv_init_stack_finished
Once the "initial stack" has finished executing, all new callbacks
are executed as the result of one of the two following options:
  1. Coming from something registered during the initial stack
     (e.g. timers, FS operations, etc.)
  2. Triggered by user input

Once Node.js enters the "loop on uv_run(..., UV_RUNONCE, ...)",
call uv_init_stack_finished to tell libuv about it.

This allows callbacks registered prior to the end of the initial
stack to identify themselves as such.