Complete implementation of origin tracking
We now flag callbacks as originating from one of the following:
  - nodejs internal
  - libuv internal
  - user
In all of the public APIs we are given a callback and a corresponding
event. We call uv__register_callback at that point, and can determine
roughly where in the code we are. This is then queried in uv_invoke_callback
to allow us to "color" trees that originate internally.

As discussed in the journal, the parentage information of
callbacks that are shunted to uv__run_pending is lost. This is
a TODO still.