CB tracking design overhaul
Change tree design a bit.

I decided that changing the parentage of a node was unwise at best.
This resulted in an after-the-fact change to the schedule, which would
make it impossible to "retrace my steps" on a subsequent run given a schedule.

Instead, I'm now maintaining dual physical/logical parentage information.
The physical tree is a faithful reproduction of the actual order in
which callbacks passed through libuv during a run.
The logical tree provides extra edges where relationships are not provided
by the physical tree. For example, the relationship of repeating timers to
themselves, or the relationships of UV__WORK_WORK to UV__WORK_DONE, is
not visible from a physical tree. However, tracking it is important,
and the logical tree gives us that.

In essence, I'm now keeping ALL the information, instead of throwing some of it away.

In support of this change, the current_callback_node is now a
per-thread map. This allows the more accurate tracking of callbacks
emanating from the thread pool.