LCBN and initial uv__register_callback code
uv__register_callback now stores an LCBN in the "context" of
the CB. The context is the handle or the req with which the CB
is associated. Handles and requests now have a map cb_type_to_lcbn
member that allows us to determine the LCBNs associated with the
next time the CB will be invoked.

uv__register_callback can only be called after uv__handle_init and uv_req_init.
A new 'magic' member for these types has been added to let us
ensure that this is the case.