LCBN: repeating CBs are part of the parent tree
The v1.0 CBN  tree included only nodes that had been executed.
The v2.0 LCBN tree includes nodes that may be executed.

In the case of ACTION nodes, they are expected to be executed.
In the case of RESPONSE nodes, they may be executed.

RESPONSE nodes are now children of the registering parent just like
ACTION nodes are.

RESPONSE nodes may be executed more than once. Each execution
must come after the previous execution of that node.
In invoke_callback, re-register the RESPONSE node as a child of the
about-to-be-executed response node.

This eliminates the segfault for ALLOC/READ CBs (which, as RESPONSEs,
are executed more than once), but it has more general ramifications.

Specifically, all nodes are now part of the same tree.
There is one root node, and it is the initial stack node.

1. This *should* make attempts to uniquely identify which LCBN
   corresponds to the next input much easier.
2. This should make it easier to determine conditional RESPONSE
   registration (conditional ACTION registration was already handled).
   On schedule variations we'll be able to slip into record or
   timeout mode more readily.

Next project: graphviz visualization of LCBN trees to verify
  these thoughts.