LCBN: Track registration time
Now we track registration time as well as CB execution start/end time.
For handles, long windows are to be expected.
For requests, long windows might be interesting to examine.

It would imply that the system is overloaded and the number of
  worker threads should be increased.
If the size of the worker threadpool were dynamic, increasing
  windows would suggest a need to increase pool, while decreasing windows
  would imply a need to decrease the pool size.
  A PID (proportional-integral-derivative) technique would work fine.