idle,check,prepare: hook into scheduler
You can now re-arrange the order of timer and idle,check,prepare CBs.

Example 1: JUST re-arranging the order of idle,check,prepare CBs
is "illegal" because it violates libuv semantics.

Example 2: Suppose idle,check,prepare CBs are registered by the AFTER_WORK_CB
associated with an FS operation. If you re-arrange the order in which the
FS operations complete, you will in turn re-arrange the order in which
the idle,check,prepare CBs are registered, which will change their execution
order.

In this case, for strict correctness you should either:
  1. re-arrange the execution order of the children on replay, or
  2. execute the first re-arrangement and then switch into RECORD mode
    (This might be a useful thing to add to the scheduler?
     But then you re-encounter non-determinism in any subsequent
     orders, which is no good.)