scheduler_advance: call before CB for nested LCBNs
Problem: "nested LCBNs" (like those added for FS operations)
are such that one CB may be executing when another executes.
In this case, the second CB triggers the assert in invoke_callback
that the CB must be next on the schedule.

Solution: Advance the scheduler prior to invoking a CB.