uv__io_poll: hook into scheduler
At the moment, any pending events in uv__io_poll will trigger
an assert. The ready_lcbns funcs have not been implemented.
However, the

 - extract ready events
 - find next context(s) and execute them
 - repair

paradigm is in place.

Note that the use of level-triggered epoll means that there's
nothing to repair. If we don't invoke a w->cb, it will just be
ready next time.