Threadpool: always allow return to uv__work_done
This change resolves benign non-determinism in RECORD mode.

Race:
Suppose there are two work items pending in the thread pool.
1. TP handles one of them and uv_async_send's.
2. Looper in uv__io_poll: heads to uv__async_event, drains the fd, and calls the ASYNC_CB
   associated with the threadpool.
3. TP handles the other one and uv_async_send's.

  Now the TP's async fd is pending again.

4. Looper goes to uv__work_done and executes both of the UV_AFTER_WORK_CBs.
5. Looper in uv__io_poll: heads to uv__async_event, drains the fd, and calls the ASYNC_CB
   associated with the threadpool.
   In uv__work_done it finds no work to do.

Solution:
In REPLAY mode, uv__work_done always uv_async_send's.
This always gives the looper thread the option to call uv__work_done.
However, it will only be called when the associated UV_ASYNC_CB is next
in the schedule.

If the associated UV_ASYNC_CB is next in the schedule, there are two options:
  1. There is no work to do (in which case it is safe to call it).
  2. Some UV_AFTER_WORK_CBs follow. These must be pending in the
     threadpool or in the loop queue (loop->wq).
     If pending in the loop queue, we will execute them immediately.
     If pending in the threadpool, we will loop in uv__work_done
     until they are available.

Consequently:
  This change will allow REPLAY to follow RECORD accurately.
  This change will not cause REPLAY to yield different results.

IMPORTANT:
This is only a fix for the race in the threadpool.
This will not fix similar races associated with other async handles.
Vanilla Node.js does not appear to make use of async handles,
though external modules might.