Improve REPLAY performance
Problem: In REPLAY mode, uv__io_poll can sometimes get "stuck"
waiting for an event or the timeout. The timeout is too long
for run_tests to survive.

Solution:
If the next event is MARKER_IO_POLL_END, we can break out of the
loop early instead of waiting for the timeout.