rescheduler: flip childless timers
This version of rescheduler can flip childless timers.

This is the simplest possible case, but should suffice
for end-to-end record+replay system testing.

timer_nondet: race between two timers
This is a good candidate for the end-to-end testing.

At the moment you have to run it like this:
  UV_SCHEDULE_MODE=REPLAY UV_SCHEDULE_FILE=raceySchedule_0 node timer_nondet.js  > /tmp/stdout 2>/tmp/stderr
because the two timers call console.log which produces a UV_WRITE_CB if stdout is going to the terminal.