Callback: properly handle nanoseconds
Was using the wrong numeric type, which resulted in rounding
on the smallest 3-5 digits of a large number.

The impact was that rescheduler's schedules would have slightly-differing
times for many callbacks as compared to the original schedule.
Thus a 'diff' on the two files would flag many changed lines, including
those that were not rescheduled.