Schedule: more reschedule improvements
Correct a few errors in the reschedule implementation.
Most notably, the registration IDs of nodes are now corrected when new nodes are inserted.
This is done by a simulating execution of the schedule.

run_tests can now reschedule and explore all of the races in var_races/*.js
over 5 iterations each.