Schedule::isValid: detect parent-after-child error
Problem:
The current rescheduler design works as follows:
  identify pivot
  for all non-pivot racy nodes:
    remove node and its descendants
    insert each removed node relative to the pivot or to the most-recently-inserted node

This assumes, however, that the nodes we are removing and re-inserting will not
end up preceding their ancestors!

This commit does not fix the problem.
However, it adds a check in Schedule::isValid to ensure that a node
is never executed *after* its descendants.

I've also introduced a test case that enables us to test solutions
to this issue. It fails at the moment, since rescheduler proposes
an invalid schedule.