rescheduler: fix infinite recursion bug
When we _relocateEvent(E), we may violate the parent-child relationships between
E and its relatives. In particular, if we move E before one of its ancestors
or dependencies, or if we move E after one of its children or dependents,
then we need to (recurse and) relocate the affected relatives.

Problem:
It might be (as in the case of threadpool work-done-work-done chains)
that the relatives we flag as needing to be moved are interrelated.
Suppose we are relocating event E and it now occurs before its parents A and B.

Suppose that, due to dependencies, A is a common parent of E and B.
Suppose that we attempt to move B (using _relocateEvent(B, pivotEvent=E)) first.
This will recursively also move A to its correct location (before B, which is now before E, so everyone's happy).
HOWEVER we would then blindly attempt to call _relocateEvent(A, pivotEvent=E).
Since E is now after A, we would then shift A *DOWN* the schedule to occur after E.
Then we would recursively update A's children (B and E), both of which must occur after A.

This would result in an infinite loop.
Interestingly, if we chose to move A first, then we would not have a problem,
since A would not drag B along with it.

The order of A and B was dependent on the order of iterating over a set(),
which meant that the behavior depended on the value of the PYTHONHASHSEED
chosen at runtime or specified in an environment variable.

As a result, this issue only manifested sometimes.
Its frequency could be tested like this:
 for seed in `seq 1 100`; do PYTHONHASHSEED=$seed timeout 3 rescheduler --schedFile record.sched --raceFile /tmp/raceFile > /tmp/resched_$seed 2>&1; echo "seed $seed rc $?"; done

Solution:
After identifying the initial list of relatives that need to be relocated in _relocateBypassedRelatives(),
relocate them one at a time, then filter out the relatives that no longer need to be attended to.

----------

This addresses the rescheduling issue described in noderacer2:a87109dadf1fd1.