scheduler: timeout-based divergence detection
We already have tree-comparison-based divergence detection.
This is carried out after completing the execution of a node,
and detects an obvious form of divergence (node has different children than recorded).

This commit adds *timeout-based* divergence detection.

In REPLAY mode, if the scheduled LCBN is never ready,
repeated calls to sched_lcbn_is_next will eventually trigger a timeout
and switch to RECORD mode. At that point, the next call to
sched_lcbn_is_next will return true and the computation will proceed.

This form of divergence can happen e.g. if you're REPLAYing a
rescheduled schedule. A registered-and-scheduled callback could be
disabled by some change in the order of a rescheduled schedule,
and thus never be ready to schedule.
We won't detect the divergence with tree-comparison-based divergence detection,
but a timeout will catch it eventually.

Note that this is a bit risky, since it might mask errors in rescheduler.
To ameliorate this concern, use the UV_SCHEDULE_MIN_N_EXECUTED_BEFORE_DIVERGENCE_ALLOWED
environment variable to put a minimum threshold before which any divergence
(whether registration-tree-based or timeout-based) is fatal.

 UV_SCHEDULE_MIN_N_EXECUTED_BEFORE_DIVERGENCE_ALLOWED value   |     effect (in REPLAY mode)
 -------------------------------------------------------------|----------------
     -1                                                       |  Divergence is fine anytime. This is the default value.
      N                                                       |  Divergence before executing N nodes is fatal

Sample values:
  -1 : don't care about divergence, e.g. when exploring a new schedule
  A, where A is the minimum value of the new exec IDs after rescheduling : any divergence prior to A is fatal, after that we're just using a best-guess schedule anyway
  N, where N is the number of exec IDs : any divergence is fatal

----------------------------------

Items changed in this commit:

1. scheduler and its callers, duh
2. Schedule: Eliminate the 'empty' loops that show up at the end of the schedule.
             These would occur when we removed the only user CBs from a loop and that
             loop was the final non-empty loop in the schedule.
             They resulted in spurious divergence; libuv won't enter the UV_RUN loop if
             there is nothing left to do.
3. rescheduler: tweak output for run_tests to easily grab the 'before and after' ID mappings
                so that it can pick a good value for UV_SCHEDULE_MIN_N_EXECUTED_BEFORE_DIVERGENCE_ALLOWED.
4. lib/timers.js: This is node source code. Explicitly call uv_timer_stop() when you call clearTimeout.
                  This is actually a fix for a bad change on my part awhile ago to give each of Node's timer objects
                  its own libuv timer.