Remove uv__ready_X_lcbns approach to scheduling
The uv__ready_X_lcbns functions allow us to ask which LCBNs were "ready"
on a handle or request. These "ready" LCBNs were used in two ways.

First, the list of "ready" LCBNs for a handle/request were passed to the scheduler
to determine which handle/request to trigger next out of a list of ready handles/requests.
Second, once a handle/request was chosen, its "ready" LCBNs could be ordered according
to the scheduler's desires.

The purpose of the approach was to allow us to generically flip the individual LCBNs.

This approach was misguided. When REPLAYing an execution, if the
exact same inputs are supplied, then the same handles/requests should always be
ready in the same order, and the same LCBNs for each handle/request should always be ready
in the same order.

While there are some special cases, like the async handle associated with the threadpool,
in general thinking about flipping handles and LCBNs tempts us to bend or violate the semantics of
libuv in unexpected ways. It's easy to lose track of what the original behavior was.

The only question we should be asking the scheduler is "What LCBN *type* will be executed next?",
which allows uv_run to loop on stages until we're ready to move on to the next stage, and allows
uv__timer_run, uv__io_poll, etc. to short-circuit when needed.

Other associated changes to functionality included in this commit:
  threadpool: Revert work and work_done
  loop-watcher: revert uv__run_X
  async: revert uv__async_event
  core: revert uv__run_closing_handles, pending