TP_FREEDOM: Fuzz epoll events (i.e. inputs)
Scheduler TP_FREEDOM will now fuzz epoll events.
This fuzzing has two aspects:
  1. Shuffle the ready epoll events. This mimics the inputs arriving
     in another order. The shuffle distance is controlled with parameter UV_SCHEDULER_IOPOLL_DEG_FREEDOM.
  2. Defer some of the ready epoll events. This mimics the inputs arriving
     at another time. The probability of deferring is controlled with parameter UV_SCHEDULER_IOPOLL_DEFER_PERC.

This fuzzing is legal because the effect is to modify the arrival time of inputs.
The system depends on epoll anyway for input awareness.
This fuzzing simulates behavior like "router/kernel was slow on some packets"
or "kernel's epoll implementation is not FIFO".
Once we decide to handle an event, the system is unperturbed except for the TP work and done queues.

Tests with http_closed_system_hard.js are promising, as described in the journal.
This fuzzing allows us to explore a much wider variety of schedules than vanilla node does
for fixed inputs.