Fix racy access to async in threadpool
Problem:
We were storing a uv_async_t in struct uv__work, which is a member of uv_work_t.
We call uv_close on this async, which asynchronously cleans up the uv_async_t.
However, the caller can delete the uv_work_t before the async gets cleaned up,
and we end up accessing delete'd memory in the uv_async_t cleanup.

Solution:
Dynamically allocate and destroy a uv_async_t for use with a struct uv__work.