Timer fuzzing: When deferring, usleep a bit
Once a timer is ready, epoll will use a timeout of 0.
Consequently, the likelihood of executing a timer "basically immediately"
(i.e. without any other interesting CBs intervening) is high, since the looper thread
will spin and probably not see new inputs on epoll.

To combat this, introduce a sleep of 5ms in uv__run_timers when deferring timers.