scheduler: Stream cb type schedule to /tmp
Since Node.js processes are long-running, we can't store info.
We have to stream it out.
Consequently, every time we run a CB, fwrite its type to a schedule file.

Use a 1MB buffer to minimize the cost of an fwrite every time we
run a callback.

By default, add the pid to the filename to deal with Node.js wrappers
that invoke Node.js programs (e.g. mocha).